Refactor code; comments
