+++ phoenix-core/src/main/java/org/apache/phoenix/calcite/PhoenixScalarFunction.java
@@ -30,10 +30,7 @@ import org.apache.calcite.rex.RexCall;
 import org.apache.calcite.schema.FunctionParameter;
 import org.apache.calcite.schema.ImplementableFunction;
 import org.apache.calcite.schema.ScalarFunction;
-import org.apache.phoenix.expression.function.FunctionExpression;
-import org.apache.phoenix.parse.FunctionParseNode.FunctionClassType;
 import org.apache.phoenix.parse.FunctionParseNode.BuiltInFunctionInfo;
-import org.apache.phoenix.parse.FunctionParseNode.BuiltInFunction;
 import org.apache.phoenix.parse.PFunction;
 import org.apache.phoenix.parse.PFunction.FunctionArgument;
 import org.apache.phoenix.schema.types.PDataType;
@@ -89,66 +86,6 @@ public class PhoenixScalarFunction implements ScalarFunction, ImplementableFunct
         this.pFunction = null;
     }
 
-    private static PDataType evaluateReturnType(Class<? extends FunctionExpression> f, List<FunctionArgument> argumentList) {
-        BuiltInFunction d = f.getAnnotation(BuiltInFunction.class);
-        try {
-            // Direct evaluation of the return type
-            FunctionExpression func = f.newInstance();
-            return func.getDataType();
-        } catch (Exception e) {
-            // should never happen
-            if (d.classType() == FunctionClassType.ALIAS || d.classType() == FunctionClassType.ABSTRACT) {
-                throw new RuntimeException();
-            }
-            // Grab the primary argument
-            assert(argumentList.size() != 0);
-            return PDataType.fromSqlTypeName(argumentList.get(0).getArgumentType());
-        }
-    }
-
-    public static List<PhoenixScalarFunction> createBuiltinFunctions(BuiltInFunctionInfo functionInfo){
-        List<List<FunctionArgument>> overloadedArgs = PhoenixSchema.overloadArguments(functionInfo.getArgs());
-        List<PhoenixScalarFunction> functionList = Lists.newArrayListWithExpectedSize(overloadedArgs.size());
-        Class<? extends FunctionExpression> clazz = functionInfo.getFunc();
-
-        try {
-            for (List<FunctionArgument> argumentList : overloadedArgs) {
-                List<FunctionParameter> parameters = Lists.newArrayListWithExpectedSize(argumentList.size());
-                PDataType returnType = evaluateReturnType(clazz, argumentList);
-
-                for (final FunctionArgument arg : argumentList) {
-                    parameters.add(
-                            new FunctionParameter() {
-                                public int getOrdinal() {
-                                    return arg.getArgPosition();
-                                }
-
-                                public String getName() {
-                                    return getArgumentName(arg.getArgPosition());
-                                }
-
-                                @SuppressWarnings("rawtypes")
-                                public RelDataType getType(RelDataTypeFactory typeFactory) {
-                                    PDataType dataType =
-                                            arg.isArrayType() ? PDataType.fromTypeId(PDataType.sqlArrayType(SchemaUtil
-                                                    .normalizeIdentifier(SchemaUtil.normalizeIdentifier(arg
-                                                            .getArgumentType())))) : PDataType.fromSqlTypeName(SchemaUtil
-                                                    .normalizeIdentifier(arg.getArgumentType()));
-                                    return typeFactory.createJavaType(dataType.getJavaClass());
-                                }
-
-                                public boolean isOptional() {
-                                    return arg.getDefaultValue() != null;
-                                }
-                            });
-                }
-                functionList.add(new PhoenixScalarFunction(functionInfo, parameters, returnType));
-            }
-        } catch (Exception e){
-            throw new RuntimeException("Builtin function " + functionInfo.getName() + " could not be registered", e);
-        }
-        return functionList;
-    }
     @Override
     public RelDataType getReturnType(RelDataTypeFactory typeFactory) {
         return typeFactory.createJavaType(returnType.getJavaClass());
